+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased (v3)
 * Bump minimum supported deployment target to iOS 16+
 * Require Xcode 16.2.0+ and Swift 5.10+
+* Add the `prefersEphemeralWebBrowserSession` property to the `POPPopupBridge` initializer, which specifies whether to request a private authentication session from the browser.
 
 ## 2.2.0  (2025-02-05)
 * Require Xcode 15.0+ and Swift 5.9+ (per [App Store requirements](https://developer.apple.com/news/?id=khzvxn8a))