+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # PopupBridge iOS Release Notes
 
-## unreleased (v3)
+## unreleased
 * Bump minimum supported deployment target to iOS 16+
 * Require Xcode 16.2.0+ and Swift 5.10+
 * Add the `prefersEphemeralWebBrowserSession` property to the `POPPopupBridge` initializer, which specifies whether to request a private authentication session from the browser.