+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # PopupBridge iOS Release Notes
 
-## unreleased
+## 3.0.0 (2025-04-01)
 * Breaking Changes
   * Bump minimum supported deployment target to iOS 16+
   * Require Xcode 16.2.0+ and Swift 5.10+